small bugfix

+++ expression_syntax.hpp
@@ -39,7 +39,7 @@ template<typename literal_type> class literal_expression_syntax final : public e
     fundamental_type get_return_type() const
     {
         if (std::is_same<literal_type, char>::value) return fundamental_type::Byte;
-        if (std::is_same<literal_type, int>::value) return fundamental_type::Byte;
+        if (std::is_same<literal_type, int>::value) return fundamental_type::Int;
         if (std::is_same<literal_type, bool>::value) return fundamental_type::Bool;
         if (std::is_same<literal_type, std::string>::value) return fundamental_type::String;
         if (std::is_same<literal_type, void>::value) return fundamental_type::Void;
